+++ src/tui/screens/index.cpp
@@ -203,7 +203,8 @@ namespace bookwyrm::tui::screen {
              * padding on the right side of it (e.g. up to and including the first char
              * in the next column, hence the magic offset).
              */
-            const auto start = col.startx + str.length() - trunc_len, end = col.startx + col.width + 2;
+            const auto start = col.startx + std::mbstowcs(nullptr, str.c_str(), str.size()) - trunc_len,
+                       end = col.startx + col.width + 2;
             for (auto x = start; x <= end; x++)
                 print(x, y, " ", attrs);
         }